Исправлены замечания
