+++ src/main/kotlin/org/tbank/managernews/client/ClientGo.kt
@@ -1,18 +1,16 @@
 package org.tbank.managernews.client
 
-import io.ktor.client.*
-import io.ktor.client.call.*
-import io.ktor.client.plugins.contentnegotiation.*
-import io.ktor.client.request.*
-import io.ktor.serialization.kotlinx.json.*
+import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.request.get
+import io.ktor.client.request.parameter
+import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import org.tbank.managernews.dto.News
-import java.io.File
-import java.nio.file.Paths
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import kotlin.io.path.exists
 
 
 class ClientGo {
@@ -52,22 +50,6 @@ class ClientGo {
             logger.info("Работа с HTTP,API завершина")
         }
     }
-
-    fun saveNews(path: String = "src/main/resources/news.csv", news: Collection<News>) {
-        val filePath = Paths.get(path)
-        require(!filePath.exists()) {
-            logger.error("Файл существует")  // кидает исключение, если файл существует
-        }
-        val csv = news.joinToString("\n") { news ->
-            "${news.rating},${news.id},${news.title},${news.place?.title ?: "Неизвестно"},${news.description},${news.publicationDate},${news.siteUrl},${news.favoritesCount},${news.commentsCount}"
-        }
-
-        try {
-            File(path).writeText(csv)
-            logger.info("Файл с данными сохранен");
-        } catch (e: Exception) {
-            throw RuntimeException(e);
-        }
-    }
 }
+
