+++ src/main/kotlin/org/tbank/managernews/dto/News.kt
@@ -1,6 +1,5 @@
 package org.tbank.managernews.dto
 
-import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalDateTime
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -24,16 +23,10 @@ data class News(
     val rating: Double
         get() = 1.0 / (1 + exp(-(favoritesCount / (commentsCount + 1)).toDouble()))
 }
+
 @Serializable
 data class Place(
     val title: String,
     val address: String
 )
-fun List<News>.getMostRatedNews(
-    count: Int,
-    period: ClosedRange<LocalDate>
-): List<News> {
-    return this.filter { news ->
-        news.publicationDate.date in period
-    }.sortedByDescending { it.rating }.take(count)
-}
+
